feat(save and return enter code): fixed unit tests
