+++ packages/appeals-service-api/src/lib/notify.js
@@ -103,9 +103,19 @@ const sendSaveAndReturnEnterCodeIntoServiceEmail = async (appeal, token) => {
   }
 };
 
+const createToken = () => {
+  const token = [];
+  for (let i = 0; i < 5; i += 1) {
+    const num = Math.floor(Math.random() * 9 + 1);
+    token.push(num);
+  }
+  return +token.join('');
+};
+
 module.exports = {
   sendSubmissionReceivedEmailToLpa,
   sendSubmissionConfirmationEmailToAppellant,
   sendSaveAndReturnContinueWithAppealEmail,
   sendSaveAndReturnEnterCodeIntoServiceEmail,
+  createToken,
 };
