+++ packages/appeals-service-api/src/services/save-and-return.service.js
@@ -3,16 +3,8 @@ const mongodb = require('../db/db');
 const {
   sendSaveAndReturnContinueWithAppealEmail,
   sendSaveAndReturnEnterCodeIntoServiceEmail,
+  createToken,
 } = require('../lib/notify');
-
-const createToken = () => {
-  const token = [];
-  for (let i = 0; i < 5; i += 1) {
-    const num = Math.floor(Math.random() * 9 + 1);
-    token.push(num);
-  }
-  return +token.join('');
-};
 
 const saveAndReturnCreateService = async (appeal) => {
   const token = createToken();
@@ -95,5 +87,4 @@ module.exports = {
   saveAndReturnTokenService,
   saveAndReturnNotifyContinue,
   saveAndReturnNotifyCode,
-  createToken,
 };
